+++ server/src/main/java/org/opensearch/transport/TcpTransport.java
@@ -778,15 +778,11 @@ public abstract class TcpTransport extends AbstractLifecycleComponent implements
     protected abstract void stopInternal();
 
     /**
-     * @deprecated use {@link #inboundMessage(TcpChannel, ProtocolInboundMessage)}
-     * Handles inbound message that has been decoded.
-     *
-     * @param channel the channel the message is from
-     * @param message the message
+     * @deprecated Use {{@link #inboundMessage(TcpChannel, InboundMessage)}} instead
      */
-    @Deprecated(since = "2.14.0", forRemoval = true)
-    public void inboundMessage(TcpChannel channel, InboundMessage message) {
-        inboundMessage(channel, (ProtocolInboundMessage) message);
+    @Deprecated
+    public void inboundMessage(TcpChannel channel, ProtocolInboundMessage message) {
+        inboundMessage(channel, (InboundMessage) message);
     }
 
     /**
@@ -795,7 +791,7 @@ public abstract class TcpTransport extends AbstractLifecycleComponent implements
      * @param channel the channel the message is from
      * @param message the message
      */
-    public void inboundMessage(TcpChannel channel, ProtocolInboundMessage message) {
+    public void inboundMessage(TcpChannel channel, InboundMessage message) {
         try {
             inboundHandler.inboundMessage(channel, message);
         } catch (Exception e) {
